Refactor extract, separate annotation function
